refactor(alert): template loading to use ParseFS instead of ParseFiles (fix #296)
This change updates the way alert templates are loaded in the
internal/alert package by replacing the use of template.ParseFiles
with template.ParseFS. This is done to make the loading of templates
more flexible and allow for more options when specifying the location
of the template files. Additionally, the change also updates the
template file path to be more consistent with the project's file
structure. This change will not affect the functionality of the program,
but it will improve maintainability and scalability in the long run.